requirements/base.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.29.0

# Bleeding edge Django
django==1.10.6

# Configuration
django-environ==0.4.1



# Images
Pillow==4.0.0

# Password storage
argon2-cffi==16.3.0

# Python-PostgreSQL Database Adapter
psycopg2==2.7

# Time zones support
pytz==2016.10

# Redis support
django-redis==4.7.0
redis>=2.10.5

# For periodical tasks
celery==4.0.2

# DRF
djangorestframework==3.6.2
django-rest-swagger==2.1.2
drf-extensions==0.3.1
